Refactor _ask.py: move Llama Index related code to _llama_index.py and add test
